+++ src/components/template/Header.tsx
@@ -18,7 +18,20 @@ import type { HeaderProps, MenuProp } from "nxs-navigation";
  * @returns Header component
  */
 const Header: React.FC<HeaderProps> = (props) => {
-  const { menu, logo, ping, language, theme, uniqueId, themeList, includeHome, layout, hideIcons, utilities, activePath } = props;
+  const {
+    menu,
+    logo,
+    ping,
+    language,
+    theme,
+    uniqueId,
+    themeList,
+    includeHome,
+    layout,
+    hideIcons,
+    utilities,
+    activePath,
+  } = props;
   const { updateMenu, onLogoClick, handleTheme, onHomeClick } = props;
   const { lightColor, errors } = useRequiredProps({ menu }, true);
   const [isActive, setActive] = useState(false);